Add /dev.html page w/ full redux dev tools
* normal page wont have store instrumented to stay lean
* reshuffled vendor deps to balance bundle sizes
