+++ client/app/scripts/stores/configureStore.js
@@ -1,5 +1,12 @@
-if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./configureStore.prod');
-} else {
-  module.exports = require('./configureStore.dev');
+import { createStore, applyMiddleware } from 'redux';
+import thunkMiddleware from 'redux-thunk';
+
+import { initialState, rootReducer } from '../reducers/root';
+
+export default function configureStore() {
+  return createStore(
+    rootReducer,
+    initialState,
+    applyMiddleware(thunkMiddleware)
+  );
 }
