--- client/app/scripts/stores/configureStore.prod.js
@@ -1,12 +1,0 @@
-import { createStore, applyMiddleware } from 'redux';
-import thunkMiddleware from 'redux-thunk';
-
-import { initialState, rootReducer } from '../reducers/root';
-
-export default function configureStore() {
-  return createStore(
-    rootReducer,
-    initialState,
-    applyMiddleware(thunkMiddleware)
-  );
-}
